Implemented PlayerList object rather than sending packets directly via Player
